added some error checking to up* functions

+++ functions/upapt.fish
@@ -1,9 +1,13 @@
 function upapt -d 'Fully update an box using APT package manager'
-    sudo apt-get -y update
-    sudo apt-get -y upgrade
-    sudo apt-get -y dist-upgrade
-    sudo apt-get -y autoremove
-    sudo apt-get -y autoclean
-    sudo apt-get check
+    if which apt-get > /dev/null ^ /dev/null
+        sudo apt-get -y update
+        sudo apt-get -y upgrade
+        sudo apt-get -y dist-upgrade
+        sudo apt-get -y autoremove
+        sudo apt-get -y autoclean
+        sudo apt-get check
+    else
+        echo 'Error: apt-get not found'
+    end
 end
